Cleanup instrumentation
