+++ instrumentation/spring-ai-client-chat-1.0.0/build.gradle
@@ -22,6 +22,6 @@ java {
 }
 
 site {
-    title 'Spring AI Client Chat'
+    title 'Spring AI Chat Completions'
     type 'AI'
 }
