+++ instrumentation/spring-ai-client-chat-1.0.0/src/main/java/llm/completions/events/LlmEvent.java
@@ -5,13 +5,13 @@
  *
  */
 
-package llm.events;
+package llm.completions.events;
 
 import com.newrelic.api.agent.NewRelic;
-import llm.models.ModelInvocation;
-import llm.models.ModelRequest;
-import llm.models.ModelResponse;
-import llm.vendor.Vendor;
+import llm.completions.models.ModelInvocation;
+import llm.completions.models.ModelRequest;
+import llm.completions.models.ModelResponse;
+import llm.completions.vendor.Vendor;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -26,23 +26,35 @@ public class LlmEvent {
     private final Map<String, Object> userLlmAttributes;
 
     // LLM event types
-    public static final String LLM_EMBEDDING = "LlmEmbedding";
+    public static final String LLM_CHAT_COMPLETION_SUMMARY = "LlmChatCompletionSummary";
+    public static final String LLM_CHAT_COMPLETION_MESSAGE = "LlmChatCompletionMessage";
 
-    // Optional LlmEvent attributes for LlmEmbedding event
+    // Optional LlmEvent attributes
     private final String spanId;
     private final String traceId;
     private final String vendor;
     private final String ingestSource;
     private final String id;
+    private final String content;
+    private final String role;
+    private final Boolean isResponse;
     private final String requestId;
     private final String responseModel;
     private final String responseOrganization;
     private final Integer responseUsageTotalTokens;
+    private final Integer responseUsagePromptTokens;
+    private final Integer responseUsageCompletionTokens;
+    private final Integer timeToFirstToken;
+    private final Integer sequence;
+    private final String completionId;
+    private final Integer responseNumberOfMessages;
     private final Float duration;
     private final Boolean error;
-    private final String input;
+    private final Float requestTemperature;
+    private final Integer requestMaxTokens;
     private final String requestModel;
     private final Integer tokenCount;
+    private final String responseChoicesFinishReason;
 
     public static class Builder {
         // Required builder parameters
@@ -65,15 +77,26 @@ public class LlmEvent {
         private String vendor = null;
         private String ingestSource = null;
         private String id = null;
+        private String content = null;
+        private String role = null;
+        private Boolean isResponse = null;
         private String requestId = null;
         private String responseModel = null;
         private String responseOrganization = null;
         private Integer responseUsageTotalTokens = null;
+        private Integer responseUsagePromptTokens = null;
+        private Integer responseUsageCompletionTokens = null;
+        private Integer timeToFirstToken = null;
+        private Integer sequence = null;
+        private String completionId = null;
+        private Integer responseNumberOfMessages = null;
         private Float duration = null;
         private Boolean error = null;
-        private String input = null;
+        private Float requestTemperature = null;
+        private Integer requestMaxTokens = null;
         private String requestModel = null;
         private Integer tokenCount = null;
+        private String responseChoicesFinishReason = null;
 
         public Builder(ModelInvocation modelInvocation) {
             userAttributes = modelInvocation.getUserAttributes();
@@ -107,6 +130,25 @@ public class LlmEvent {
             return this;
         }
 
+        public Builder content(String message) {
+            content = message;
+            return this;
+        }
+
+        public Builder role(boolean isUser) {
+            if (isUser) {
+                role = "user";
+            } else {
+                role = "assistant";
+            }
+            return this;
+        }
+
+        public Builder isResponse(boolean isUser) {
+            isResponse = !isUser;
+            return this;
+        }
+
         public Builder requestId() {
             requestId = modelResponse.getRequestId();
             return this;
@@ -127,6 +169,36 @@ public class LlmEvent {
             return this;
         }
 
+        public Builder responseUsagePromptTokens() {
+            responseUsagePromptTokens = modelResponse.getResponseUsagePromptTokens();
+            return this;
+        }
+
+        public Builder responseUsageCompletionTokens() {
+            responseUsageCompletionTokens = modelResponse.getResponseUsageCompletionTokens();
+            return this;
+        }
+
+        public Builder timeToFirstToken(int time) {
+            timeToFirstToken = modelResponse.getTimeToFirstToken();
+            return this;
+        }
+
+        public Builder sequence(int eventSequence) {
+            sequence = eventSequence;
+            return this;
+        }
+
+        public Builder completionId() {
+            completionId = modelResponse.getLlmChatCompletionSummaryId();
+            return this;
+        }
+
+        public Builder responseNumberOfMessages(int numberOfMessages) {
+            responseNumberOfMessages = numberOfMessages;
+            return this;
+        }
+
         public Builder duration(float callDuration) {
             duration = callDuration;
             return this;
@@ -137,8 +209,13 @@ public class LlmEvent {
             return this;
         }
 
-        public Builder input(int index) {
-            input = modelRequest.getInputText(index);
+        public Builder requestTemperature() {
+            requestTemperature = modelRequest.getTemperature();
+            return this;
+        }
+
+        public Builder requestMaxTokens() {
+            requestMaxTokens = modelRequest.getMaxTokensToSample();
             return this;
         }
 
@@ -149,6 +226,11 @@ public class LlmEvent {
 
         public Builder tokenCount(Integer count) {
             tokenCount = count;
+            return this;
+        }
+
+        public Builder responseChoicesFinishReason() {
+            responseChoicesFinishReason = modelResponse.getStopReason();
             return this;
         }
 
@@ -190,6 +272,21 @@ public class LlmEvent {
             eventAttributes.put("id", id);
         }
 
+        content = builder.content;
+        if (isAiMonitoringRecordContentEnabled() && content != null && !content.isEmpty()) {
+            eventAttributes.put("content", content);
+        }
+
+        role = builder.role;
+        if (role != null && !role.isEmpty()) {
+            eventAttributes.put("role", role);
+        }
+
+        isResponse = builder.isResponse;
+        if (isResponse != null) {
+            eventAttributes.put("is_response", isResponse);
+        }
+
         requestId = builder.requestId;
         if (requestId != null && !requestId.isEmpty()) {
             eventAttributes.put("request_id", requestId);
@@ -210,6 +307,36 @@ public class LlmEvent {
             eventAttributes.put("response.usage.total_tokens", responseUsageTotalTokens);
         }
 
+        responseUsagePromptTokens = builder.responseUsagePromptTokens;
+        if (responseUsagePromptTokens != null && responseUsagePromptTokens >= 0) {
+            eventAttributes.put("response.usage.prompt_tokens", responseUsagePromptTokens);
+        }
+
+        responseUsageCompletionTokens = builder.responseUsageCompletionTokens;
+        if (responseUsageCompletionTokens != null && responseUsageCompletionTokens >= 0) {
+            eventAttributes.put("response.usage.completion_tokens", responseUsageCompletionTokens);
+        }
+
+        timeToFirstToken = builder.timeToFirstToken;
+        if (timeToFirstToken != null && timeToFirstToken >= 0) {
+            eventAttributes.put("time_to_first_token", timeToFirstToken);
+        }
+
+        sequence = builder.sequence;
+        if (sequence != null && sequence >= 0) {
+            eventAttributes.put("sequence", sequence);
+        }
+
+        completionId = builder.completionId;
+        if (completionId != null && !completionId.isEmpty()) {
+            eventAttributes.put("completion_id", completionId);
+        }
+
+        responseNumberOfMessages = builder.responseNumberOfMessages;
+        if (responseNumberOfMessages != null && responseNumberOfMessages >= 0) {
+            eventAttributes.put("response.number_of_messages", responseNumberOfMessages);
+        }
+
         duration = builder.duration;
         if (duration != null && duration >= 0) {
             eventAttributes.put("duration", duration);
@@ -220,9 +347,14 @@ public class LlmEvent {
             eventAttributes.put("error", true);
         }
 
-        input = builder.input;
-        if (isAiMonitoringRecordContentEnabled() && input != null && !input.isEmpty()) {
-            eventAttributes.put("input", input);
+        requestTemperature = builder.requestTemperature;
+        if (requestTemperature != null && requestTemperature >= 0) {
+            eventAttributes.put("request.temperature", requestTemperature);
+        }
+
+        requestMaxTokens = builder.requestMaxTokens;
+        if (requestMaxTokens != null && requestMaxTokens > 0) {
+            eventAttributes.put("request.max_tokens", requestMaxTokens);
         }
 
         requestModel = builder.requestModel;
@@ -233,6 +365,11 @@ public class LlmEvent {
         tokenCount = builder.tokenCount;
         if (tokenCount != null && tokenCount > 0) {
             eventAttributes.put("token_count", tokenCount);
+        }
+
+        responseChoicesFinishReason = builder.responseChoicesFinishReason;
+        if (responseChoicesFinishReason != null && !responseChoicesFinishReason.isEmpty()) {
+            eventAttributes.put("response.choices.finish_reason", responseChoicesFinishReason);
         }
     }
 
@@ -258,9 +395,16 @@ public class LlmEvent {
     }
 
     /**
-     * Record a LlmEmbedding custom event
+     * Record a LlmChatCompletionMessage custom event
      */
-    public void recordLlmEmbeddingEvent() {
-        NewRelic.getAgent().getInsights().recordCustomEvent(LLM_EMBEDDING, eventAttributes);
+    public void recordLlmChatCompletionMessageEvent() {
+        NewRelic.getAgent().getInsights().recordCustomEvent(LLM_CHAT_COMPLETION_MESSAGE, eventAttributes);
+    }
+
+    /**
+     * Record a LlmChatCompletionSummary custom event
+     */
+    public void recordLlmChatCompletionSummaryEvent() {
+        NewRelic.getAgent().getInsights().recordCustomEvent(LLM_CHAT_COMPLETION_SUMMARY, eventAttributes);
     }
 }
