+++ instrumentation/spring-ai-client-chat-1.0.0/src/main/java/llm/completions/models/ModelResponse.java
@@ -5,16 +5,11 @@
  *
  */
 
-package llm.models;
-
-import com.newrelic.api.agent.NewRelic;
-
-import java.util.logging.Level;
+package llm.completions.models;
 
 public interface ModelResponse {
-    // Operation types
+    // Operation type
     String COMPLETION = "completion";
-    String EMBEDDING = "embedding";
 
     /**
      * Get the response message, potentially from a specific array index
@@ -61,13 +56,6 @@ public interface ModelResponse {
     String getLlmChatCompletionSummaryId();
 
     /**
-     * Get the ID for the associated LlmEmbedding event.
-     *
-     * @return String representing the ID for the associated LlmEmbedding event
-     */
-    String getLlmEmbeddingId();
-
-    /**
      * Determine whether the response resulted in an error or not.
      *
      * @return boolean true when the LLM response is an error, false when the response was successful
@@ -87,20 +75,6 @@ public interface ModelResponse {
      * @return String representing the response status text
      */
     String getStatusText();
-
-    /**
-     * Log when a parsing error occurs.
-     *
-     * @param e                Exception encountered when parsing the response
-     * @param fieldBeingParsed field that was being parsed
-     */
-    static void logParsingFailure(Exception e, String fieldBeingParsed) {
-        if (e != null) {
-            NewRelic.getAgent().getLogger().log(Level.FINEST, e, "AIM: Error parsing " + fieldBeingParsed + " from ModelResponse");
-        } else {
-            NewRelic.getAgent().getLogger().log(Level.FINEST, "AIM: Unable to parse empty/null " + fieldBeingParsed + " from ModelResponse");
-        }
-    }
 
     /**
      * Get the response model ID.
