+++ instrumentation/spring-ai-client-chat-1.0.0/src/main/java/llm/completions/models/springai/README.md
@@ -1,0 +1,6 @@
+# Spring AI
+
+Examples of the request/response bodies for models that have been tested and verified to work. The instrumentation should continue to correctly process new
+models as long as they match the model naming prefixes in `llm.completions.models.SupportedModels` and the request/response structure stays the same as the examples listed
+here.
+
