+++ instrumentation/spring-ai-client-chat-1.0.0/src/main/java/llm/completions/models/springai/SpringAiModelRequest.java
@@ -5,16 +5,17 @@
  *
  */
 
-package llm.models.springai;
+package llm.completions.models.springai;
 
 import com.newrelic.api.agent.NewRelic;
-import llm.models.ModelRequest;
+import llm.completions.models.ModelRequest;
 import org.springframework.ai.chat.client.ChatClientRequest;
 import org.springframework.ai.chat.messages.MessageType;
 import org.springframework.ai.chat.messages.UserMessage;
 import org.springframework.ai.chat.prompt.ChatOptions;
 import org.springframework.ai.chat.prompt.Prompt;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.logging.Level;
 
@@ -24,7 +25,7 @@ import java.util.logging.Level;
  */
 public class SpringAiModelRequest implements ModelRequest {
     private static final String USER = "user";
-    private String promptUserMessage = "";
+    private List<UserMessage> promptUserMessages = new ArrayList<>();
     private String modelId = "";
     private Integer maxTokensToSample = 0;
     private float temperature = 0;
@@ -50,22 +51,21 @@ public class SpringAiModelRequest implements ModelRequest {
                     }
                 }
 
-                List<UserMessage> userMessages = prompt.getUserMessages();
-                if (userMessages != null) {
-                    numberOfRequestMessages = userMessages.size();
+                promptUserMessages.addAll(prompt.getUserMessages());
+                if (promptUserMessages != null) {
+                    numberOfRequestMessages = promptUserMessages.size();
                 }
 
                 UserMessage userMessage = prompt.getUserMessage();
                 if (userMessage != null) {
-                    promptUserMessage = userMessage.getText();
                     MessageType messageType = userMessage.getMessageType();
                     if (messageType != null) {
                         messageTypeValue = messageType.getValue();
                     }
                 }
-            } else {
-                NewRelic.getAgent().getLogger().log(Level.FINEST, "AIM: Received null SpringAI ChatClientRequest");
             }
+        } else {
+            NewRelic.getAgent().getLogger().log(Level.FINEST, "AIM: Received null SpringAI ChatClientRequest");
         }
     }
 
@@ -86,17 +86,11 @@ public class SpringAiModelRequest implements ModelRequest {
 
     @Override
     public String getRequestMessage(int index) {
-        return promptUserMessage;
-    }
-
-    @Override
-    public String getInputText(int index) {
-        return ""; // TODO debug embedding
-    }
-
-    @Override
-    public int getNumberOfInputTextMessages() {
-        return 0; // TODO debug embedding
+        UserMessage userMessage = promptUserMessages.get(index);
+        if (userMessage != null) {
+            return userMessage.getText();
+        }
+        return "";
     }
 
     @Override
