+++ instrumentation/spring-ai-model-1.0.0/build.gradle
@@ -4,7 +4,7 @@ jar {
 
 dependencies {
     implementation(project(":agent-bridge"))
-    implementation 'org.springframework.ai:spring-ai-model:1.0.0' // compile against version 1.0.0 in final implementation
+    implementation 'org.springframework.ai:spring-ai-model:1.0.0'
 
     testImplementation 'org.springframework.ai:spring-ai-model:1.0.0'
 //    testImplementation 'org.mockito:mockito-inline:4.11.0'
@@ -22,6 +22,6 @@ java {
 }
 
 site {
-    title 'Spring AI Embedding Model'
+    title 'Spring AI Embeddings'
     type 'AI'
 }
