+++ instrumentation/spring-ai-model-1.0.0/src/main/java/llm/embeddings/models/ModelResponse.java
@@ -5,7 +5,7 @@
  *
  */
 
-package llm.models;
+package llm.embeddings.models;
 
 public interface ModelResponse {
     // Operation type
