+++ instrumentation/spring-ai-model-1.0.0/src/main/java/llm/embeddings/models/SupportedModels.java
@@ -1,0 +1,18 @@
+/*
+ *
+ *  * Copyright 2026 New Relic Corporation. All rights reserved.
+ *  * SPDX-License-Identifier: Apache-2.0
+ *
+ */
+
+package llm.embeddings.models;
+
+/**
+ * Prefixes for supported models. As long as the model ID for an invoked LLM model contains
+ * one of these prefixes the instrumentation should attempt to process the request/response.
+ * <p>
+ * See the README for each model in llm.embeddings.models.* for more details on supported models.
+ */
+public class SupportedModels {
+//    public static final String OPEN_AI = "openai";
+}
