+++ instrumentation/spring-ai-model-1.0.0/src/test/java/software/amazon/awssdk/services/bedrockruntime/BedrockRuntimeClient_InstrumentationTest.java
@@ -14,7 +14,7 @@
 //import com.newrelic.agent.introspec.TracedMetricData;
 //import com.newrelic.api.agent.NewRelic;
 //import com.newrelic.api.agent.Trace;
-//import llm.models.ModelResponse;
+//import llm.embeddings.models.ModelResponse;
 //import org.json.JSONObject;
 //import org.junit.Before;
 //import org.junit.Test;
